# Backend/docs/api-contract.yaml
openapi: 3.0.3
info:
  title: HSE C# Gamification Backend API
  version: 1.0.0
  description: Frontend-oriented API contract for the current backend.
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TokenRead:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: bearer }
    UserCreate:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
        full_name: { type: string, nullable: true }
    UserRead:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        full_name: { type: string, nullable: true }
        role: { type: string, enum: [user, teacher, admin] }
...
    AnswerCreate:
      type: object
      required: [test_id, question_id, answer_payload]
      properties:
        test_id: { type: integer }
        attempt_id: { type: integer, nullable: true }
        question_id: { type: integer }
        answer_payload:
          type: string
          description: choice id as string for closed questions, free text for open questions
    AnswerRead:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        test_id: { type: integer }
        attempt_id: { type: integer, nullable: true }
        question_id: { type: integer }
        answer_payload: { type: string }
        score: { type: number, format: float, nullable: true }
        graded_by: { type: integer, nullable: true }
        graded_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
...
    GradeRequest:
      type: object
      required: [score]
      properties:
        score: { type: number, format: float }
        comment: { type: string, nullable: true }
...
paths:
...
  /api/v1/auth/login:
    post:
      security: []
      summary: JSON login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenRead' }
  /api/v1/auth/register:
    post:
      security: []
      summary: Self-register as user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserRead' }
...
  /api/v1/answers:
    post:
      summary: Submit answer
      description: Backend resolves or auto-creates active attempt if attempt_id is omitted.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnswerCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnswerRead' }
        '409': { description: Deadline passed, time limit exceeded, or test already completed }
...
  /api/v1/answers/{answer_id}/grade:
    post:
      summary: Manual grade for open answer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GradeRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnswerRead' }
...